HADOOP-18724. Pull test up.
move the test with a floating point len from hadoop-aws to hadoop-common,
so verify that all filesystems can cope with it.
